switched tests in ObjectComparsion to FluentAssertions and refactored tests
